pyproject.toml
[project]
name = "ks-cli"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "certifi>=2026.1.4",
    "click>=8.3.1",
    "httpx>=0.28.1",
    "ksapi>=1.25.0",
    "pyjwt>=2.11.0",
    "rich>=14.3.3",
]

[project.scripts]
kscli = "kscli.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/kscli"]

[dependency-groups]
dev = ["basedpyright>=1.38.1", "pytest>=9.0.2", "ruff>=0.15.2"]

# Pyright configuration - strict type checking
[tool.basedpyright]
include = ["src/"]
pythonVersion = "3.12"
typeCheckingMode = "standard"
useLibraryCodeForTypes = true
# Add the project root to the Python path so imports work correctly
reportGeneralTypeIssues = "error"
reportOptionalMemberAccess = "error"
reportOptionalCall = "error"
reportOptionalIterable = "error"
reportOptionalContextManager = "error"
reportOptionalOperand = "error"
reportTypedDictNotRequiredAccess = "error"
reportPrivateImportUsage = "error"
reportConstantRedefinition = "error"
reportIncompatibleMethodOverride = "error"
reportIncompatibleVariableOverride = "error"
reportInconsistentConstructor = "error"
reportOverlappingOverload = "error"
reportMissingTypeArgument = "error"
reportAny = "none"
reportMissingImports = "error"
reportMissingTypeStubs = "error"
